Renamed the added test from Mongo* to MongoDB*
