+++ data-index/data-index-service/data-index-service-common/src/main/resources/META-INF/microprofile-config.properties
@@ -53,6 +53,7 @@ mp.messaging.incoming.kogito-jobs-events.path=/jobs
 %kafka-events-support.mp.messaging.incoming.kogito-usertaskinstances-events.connector=smallrye-kafka
 %kafka-events-support.mp.messaging.incoming.kogito-jobs-events.connector=smallrye-kafka
 %kafka-events-support.quarkus.kafka.bootstrap-servers=localhost:9092
+%kafka-events-support.kafka.bootstrap.servers=localhost:9092
 
 mp.messaging.incoming.kogito-processinstances-events.topic=kogito-processinstances-events
 mp.messaging.incoming.kogito-processinstances-events.group.id=kogito-data-index-processinstances
