+++ data-index/data-index-service/data-index-service-mongodb/src/test/java/org/kie/kogito/index/service/messaging/MongoDBDomainMessagingHttpConsumerIT.java
@@ -17,17 +17,17 @@
 package org.kie.kogito.index.service.messaging;
 
 import org.kie.kogito.index.TestUtils;
-import org.kie.kogito.index.service.test.KafkaMessageTestProfile;
+import org.kie.kogito.index.service.test.InMemoryMessageTestProfile;
 
 import io.quarkus.test.junit.QuarkusTest;
 import io.quarkus.test.junit.TestProfile;
 
 @QuarkusTest
-@TestProfile(KafkaMessageTestProfile.class)
-class MongoDomainMessagingKafkaConsumerIT extends AbstractDomainMessagingKafkaConsumerIT {
+@TestProfile(InMemoryMessageTestProfile.class)
+class MongoDBDomainMessagingHttpConsumerIT extends AbstractDomainMessagingHttpConsumerIT {
 
     @Override
     protected String getTestProtobufFileContent() throws Exception {
-        return TestUtils.readFileContent("travels-mongo.proto");
+        return TestUtils.getTravelsProtoBufferFile();
     }
 }
