+++ data-index/data-index-service/data-index-service-mongodb/src/test/java/org/kie/kogito/index/service/messaging/MongoDBDomainMessagingKafkaConsumerIT.java
@@ -16,6 +16,7 @@
 
 package org.kie.kogito.index.service.messaging;
 
+import org.kie.kogito.index.TestUtils;
 import org.kie.kogito.index.service.test.KafkaMessageTestProfile;
 
 import io.quarkus.test.junit.QuarkusTest;
@@ -23,6 +24,10 @@ import io.quarkus.test.junit.TestProfile;
 
 @QuarkusTest
 @TestProfile(KafkaMessageTestProfile.class)
-class MongoMessagingKafkaConsumerIT extends AbstractMessagingKafkaConsumerIT {
+class MongoDBDomainMessagingKafkaConsumerIT extends AbstractDomainMessagingKafkaConsumerIT {
 
+    @Override
+    protected String getTestProtobufFileContent() throws Exception {
+        return TestUtils.readFileContent("travels-mongo.proto");
+    }
 }
